feat: support stateful git cloning
this adds a new config option to enable downloading the firefox source
through a git clone instead of downloading and unpacking a tarball.

the benefit of this approach is that with a full depth clone, resolving
patches is considerably easier as git actually has the previous context
of the file.

it also unlocks incremental updates, currently when bumping the firefox
version, the engine directory is completely removed and downloaded again
which also deletes a lot of the caching the firefox build system does.

there are obviously risks with incremental updates but if anything does
go wrong, you can just delete the engine dir and download it again.

the intention with this is not necessarily to completely enable stateful
work on the engine repo itself, as in some cases we may discard local
changes.